+++ ArabClub_APIs/users/models.py
@@ -2,6 +2,7 @@ import re
 
 from django.contrib.auth.models import BaseUserManager, AbstractBaseUser
 from django.db import models
+from django.urls import reverse
 from django.utils.translation import gettext_lazy as _
 
 
@@ -17,14 +18,13 @@ class CustomUserManager(BaseUserManager):
         """
 
         if not email:
-            raise ValueError(_('The Email must be set.'))
+            raise ValueError(_("The Email must be set."))
 
         is_valid_username = re.search(
-            r'^[a-zA](?=[a-zA-Z0-9._]{8,20}$)(?!.*[_.]{2})[^_.].*[^_.]$',
-            username
+            r"^[a-zA](?=[a-zA-Z0-9._]{8,20}$)(?!.*[_.]{2})[^_.].*[^_.]$", username
         )
         if not is_valid_username:
-            raise ValueError(_('Enter Valid Username'))
+            raise ValueError(_("Enter Valid Username"))
 
         user = self.model(
             email=self.normalize_email(email),
@@ -52,9 +52,8 @@ class CustomUserManager(BaseUserManager):
 
 
 class User(AbstractBaseUser):
-    email = models.EmailField(verbose_name='Email', max_length=255, unique=True)
-    username = models.CharField(verbose_name='Username', max_length=50,
-                                unique=True)
+    email = models.EmailField(verbose_name="Email", max_length=255, unique=True)
+    username = models.CharField(verbose_name="Username", max_length=50, unique=True)
     date_of_birth = models.DateField()
 
     is_active = models.BooleanField(default=True)
@@ -62,8 +61,8 @@ class User(AbstractBaseUser):
 
     objects = CustomUserManager()
 
-    USERNAME_FIELD = 'username'
-    REQUIRED_FIELDS = ['email', 'date_of_birth']
+    USERNAME_FIELD = "username"
+    REQUIRED_FIELDS = ["email", "date_of_birth"]
 
     def __str__(self):
         return self.username
@@ -85,32 +84,39 @@ class User(AbstractBaseUser):
 
 
 class FirstNameAndLastName(models.Model):
-    first_name = models.CharField(max_length=50, verbose_name='First Name')
-    last_name = models.CharField(max_length=50, verbose_name='Last Name')
-    user = models.OneToOneField(User, primary_key=True,
-                                on_delete=models.CASCADE)
+    first_name = models.CharField(max_length=50, verbose_name="First Name")
+    last_name = models.CharField(max_length=50, verbose_name="Last Name")
+    user = models.OneToOneField(
+        User, related_name="name", primary_key=True, on_delete=models.CASCADE
+    )
 
     def __str__(self):
         return self.first_name
 
     @property
     def get_full_name(self):
-        return f'{self.first_name} {self.last_name}'
+        return f"{self.first_name} {self.last_name}"
 
 
 class Bio(models.Model):
-    bio = models.TextField(max_length=255, verbose_name='Bio')
-    user = models.OneToOneField(User, primary_key=True,
-                                on_delete=models.CASCADE)
+    bio = models.TextField(max_length=255, verbose_name="Bio")
+    user = models.OneToOneField(
+        User, related_name="bio", primary_key=True, on_delete=models.CASCADE
+    )
 
     def __str__(self):
         return self.bio
 
+    def get_absolute_url(self):
+        user = User.objects.get(pk=self.id)
+        return reverse("bio", args=[str(user)])
+
 
 class Phone(models.Model):
-    phone = models.CharField(max_length=15, verbose_name='Phone Number')
-    user = models.OneToOneField(User, primary_key=True,
-                                on_delete=models.CASCADE)
+    phone = models.CharField(max_length=15, verbose_name="Phone Number")
+    user = models.OneToOneField(
+        User, related_name="phone", primary_key=True, on_delete=models.CASCADE
+    )
 
     def __str__(self):
         return self.phone
@@ -118,8 +124,9 @@ class Phone(models.Model):
 
 class GitHubAccount(models.Model):
     url = models.URLField()
-    user = models.OneToOneField(User, primary_key=True,
-                                on_delete=models.CASCADE)
+    user = models.OneToOneField(
+        User, related_name="github_url", primary_key=True, on_delete=models.CASCADE
+    )
 
     def __str__(self):
         return self.url
@@ -127,25 +134,28 @@ class GitHubAccount(models.Model):
 
 class Skills(models.Model):
     skill_name = models.CharField(max_length=150)
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
+    user = models.OneToOneField(
+        User, related_name="skills", primary_key=True, on_delete=models.CASCADE
+    )
 
     def __str__(self):
         return self.skill_name
 
 
 class Address(models.Model):
-    country = models.CharField(max_length=50, verbose_name='Country')
-    city = models.CharField(max_length=50, verbose_name='City')
-    street_name = models.CharField(max_length=150, verbose_name='Street Name')
-    user = models.OneToOneField(User, primary_key=True,
-                                on_delete=models.CASCADE)
+    country = models.CharField(max_length=50, verbose_name="Country")
+    city = models.CharField(max_length=50, verbose_name="City")
+    street_name = models.CharField(max_length=150, verbose_name="Street Name")
+    user = models.OneToOneField(
+        User, related_name="address", primary_key=True, on_delete=models.CASCADE
+    )
 
     def __str__(self):
         return self.country
 
     @property
     def get_full_address(self):
-        return f'{self.street_name}, {self.city}, {self.country}'
+        return f"{self.street_name}, {self.city}, {self.country}"
 
     @property
     def get_country(self):