+++ ArabClub_APIs/users/serializers.py
@@ -1,53 +1,31 @@
+import re
+
 from django.contrib.auth import get_user_model
 from rest_framework import serializers
 
 from users.models import (
-    FirstNameAndLastName, Phone, Bio, Address, GitHubAccount, Skills
+    FirstNameAndLastName,
+    Phone,
+    Bio,
+    Address,
+    GitHubAccount,
+    Skills,
 )
 
 
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = get_user_model()
-        fields = ['id', 'username', 'email', 'date_of_birth']
-
-
-class UpdateUserSerializer(serializers.ModelSerializer):
-    username = serializers.CharField(required=False)
-    email = serializers.EmailField(required=False)
-    date_of_birth = serializers.DateField(required=False)
-
-    class Meta:
-        model = get_user_model()
-        fields = ['id', 'username', 'email', 'date_of_birth']
-
-    def update(self, instance, validate_date):
-        instance.username = validate_date.get('username', instance.username)
-        instance.email = validate_date.get('email', instance.email)
-        instance.date_of_birth = validate_date.get('date_of_birth',
-                                                   instance.date_of_birth)
-        instance.save()
-        return instance
-
-    @classmethod
-    def get_model(cls):
-        return cls.Meta.model
-
-
-class UpdateNamesSerializer(serializers.ModelSerializer):
+# User First name and last name serializers with custom update data method
+class NameSerializer(serializers.ModelSerializer):
     first_name = serializers.CharField(max_length=50, required=True)
     last_name = serializers.CharField(max_length=50, required=True)
-    user_id = serializers.IntegerField(required=True)
 
     class Meta:
         model = FirstNameAndLastName
-        fields = ['first_name', 'last_name', 'user_id']
+        fields = ["first_name", "last_name"]
 
     def update(self, instance, validated_data):
-        instance.first_name = validated_data.get('first_name',
-                                                 instance.first_name)
-        instance.last_name = validated_data.get('last_name', instance.last_name)
-        instance.user_id = validated_data.get('user_id', instance.user_id)
+        instance.first_name = validated_data.get("first_name", instance.first_name)
+        instance.last_name = validated_data.get("last_name", instance.last_name)
+        instance.user_id = validated_data.get("user_id", instance.user_id)
         instance.save()
         return instance
 
@@ -57,25 +35,26 @@ class UpdateNamesSerializer(serializers.ModelSerializer):
         return obj
 
 
+# User Bio serializers with custom update data method
 class BioSerializer(serializers.ModelSerializer):
     class Meta:
         model = Bio
-        fields = ['bio', 'user_id']
+        fields = ["bio"]
 
     def update(self, instance, validated_data):
-        instance.bio = validated_data.get('bio', instance.bio)
+        instance.bio = validated_data.get("bio", instance.bio)
         instance.save()
         return instance
 
 
+# User skills serializers
 class SkillsSerializer(serializers.ModelSerializer):
-    id = serializers.IntegerField(read_only=True)
-
     class Meta:
         model = Skills
-        fields = ['id', 'skill_name', 'user_id']
+        fields = ["skill_name"]
 
 
+# User address serializers with custom update data method
 class AddressSerializer(serializers.ModelSerializer):
     country = serializers.CharField(max_length=50, required=False)
     city = serializers.CharField(max_length=50, required=False)
@@ -83,131 +62,142 @@ class AddressSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = Address
-        fields = ['country', 'city', 'street_name', 'user_id']
+        fields = ["country", "city", "street_name"]
 
     def update(self, instance, validated_data):
-        instance.country = validated_data.get('country', instance.country)
-        instance.city = validated_data.get('city', instance.city)
-        instance.street_name = validated_data.get('street_name',
-                                                  instance.street_name)
+        instance.country = validated_data.get("country", instance.country)
+        instance.city = validated_data.get("city", instance.city)
+        instance.street_name = validated_data.get("street_name", instance.street_name)
         instance.save()
         return instance
 
 
+# User GitHub Account serializers
 class GitHubSerializer(serializers.ModelSerializer):
     class Meta:
         model = GitHubAccount
-        fields = ['url', 'user_id']
-
-    serializer_url_field = 'url'
+        fields = ["url"]
 
 
+# User phone number serializers
 class PhoneSerializer(serializers.ModelSerializer):
     class Meta:
         model = Phone
-        fields = ['phone', 'user_id']
+        fields = ["phone"]
+
+    def validate_phone(self, phone):
+        pattern = r"[0-9]{11,15}"
+        re.compile(pattern)
+        if re.fullmatch(pattern, phone):
+            return phone
+        raise serializers.ValidationError(
+            "Please enter your phone number " "correctly."
+        )
+
+    def update(self, instance, validated_data):
+        instance.phone = validated_data.get("phone", instance.phone)
+        instance.save()
+        return instance
 
 
-class CustomSerializerUser:
-    """
-    Get all data for users and set values to list
-    """
+class UserSerializer(serializers.ModelSerializer):
+    bio = BioSerializer(required=False)
+    phone = PhoneSerializer(required=False)
+    skills = SkillsSerializer(required=False)
+    name = NameSerializer(required=False)
+    address = AddressSerializer(required=False)
+    username = serializers.CharField(required=False)
+    email = serializers.EmailField(required=False)
+    date_of_birth = serializers.DateField(required=False)
+    github_url = GitHubSerializer(required=False)
 
-    def __init__(self, obj):
-        self.obj = obj
-        self.data = []
-        self.index = -1
-        self.id = 0
-        self.err = False
+    class Meta:
+        model = get_user_model()
+        fields = [
+            "id",
+            "name",
+            "username",
+            "email",
+            "date_of_birth",
+            "bio",
+            "skills",
+            "github_url",
+            "phone",
+            "address",
+        ]
+
+    def validate_username(self, username):
+        pattern = re.compile(
+            r"^[a-zA](?=[a-zA-Z0-9._]{8,20}$)(?!.*[_.]{2})[^_.].*[^_.]$"
+        )
+
+        if re.fullmatch(pattern, username):
+            return username
+
+        raise serializers.ValidationError("Enter valid username ex. xxxx.xxxx")
+
+    def update(self, instance, validated_data):
+        instance.username = validated_data.get("username", instance.username)
+        instance.email = validated_data.get("email", instance.email)
+        instance.date_of_birth = validated_data.get(
+            "date_of_birth", instance.date_of_birth
+        )
+        self.update_name(instance, validated_data)
+        self.update_bio(instance, validated_data)
+        self.update_skills(instance, validated_data)
+        self.update_address(instance, validated_data)
+        self.update_phone(instance, validated_data)
+        self.update_github_url(instance, validated_data)
+        instance.save()
+        return instance
+
+    def update_github_url(self, instance, validated_data):
         try:
-            self.get_basic_data_for_many()
-        except Exception as e:
-            self.get_basic_data_for_single()
-
-    def get_basic_data_for_single(self) -> list:
-        try:
-            self.id = self.obj.id
-        except AttributeError:
-            self.err = '[Errors 01] 404 user not found'
-            return self.err
-
-        self.index += 1
-        self.data.append({'id': self.id, 'username': self.obj.username})
-        bio = self.bio
-        self.data[self.index]['birthdate'] = self.obj.date_of_birth
-        self.get_person_info()
-        self.get_address()
-        self.get_skills()
-        self.get_phone()
-        self.get_github()
-
-        return self.data
-
-    def get_basic_data_for_many(self):
-        for user in self.obj:
-            self.id = user.id
-            self.index += 1
-            self.data.append({'id': self.id, 'username': user.username})
-            bio = self.bio
-            self.data[self.index]['birthdate'] = user.date_of_birth
-            self.get_person_info()
-            self.get_address()
-            self.get_skills()
-            self.get_phone()
-            self.data[self.index]['person_info']['connect']['email'] = \
-                user.email
-
-            self.get_github()
-
-        return self.data
-
-    @property
-    def bio(self):
-        try:
-            bio = Bio.objects.get(user_id=self.id)
-            self.data[self.index]['bio'] = bio.bio
-        except Bio.DoesNotExist:
+            github_data = validated_data.pop("github_url")
+            if not GitHubAccount.objects.update(user_id=instance.id, **github_data):
+                GitHubAccount.objects.create(user_id=instance.id)
+        except KeyError:
             pass
 
-    def get_person_info(self):
+    def update_phone(self, instance, validated_data):
         try:
-            names = FirstNameAndLastName.objects.get(user_id=self.id)
-            self.data[self.index]['person_info'] = {
-                'first_name': names.first_name,
-                'last_name': names.last_name,
-            }
-        except FirstNameAndLastName.DoesNotExist:
+            phone_data = validated_data.pop("phone")
+            if not Phone.objects.update(user_id=instance.id, **phone_data):
+                Phone.objects.create(user_id=instance.id, **phone_data)
+        except KeyError:
             pass
 
-    def get_github(self):
+    def update_address(self, instance, validated_data):
         try:
-            github = GitHubAccount.objects.get(user_id=self.id)
-            self.data[self.index]['github'] = github.url
-        except GitHubAccount.DoesNotExist:
+            address_data = validated_data.pop("address")
+            if not Address.objects.update(user_id=instance.id, **address_data):
+                Address.objects.create(user_id=instance.id, **address_data)
+        except KeyError:
             pass
 
-    def get_phone(self):
+    def update_skills(self, instance, validated_data):
         try:
-            phone = Phone.objects.get(user_id=self.id)
-            self.data[self.index]['person_info']['connect'] = {
-                'phone': phone.phone}
-        except Phone.DoesNotExist:
+            skill_data = validated_data.pop("skills")
+            if not Skills.objects.update(user_id=instance.id, **skill_data):
+                Skills.objects.create(user_id=instance.id, **skill_data)
+        except KeyError:
             pass
 
-    def get_skills(self):
+    def update_bio(self, instance, validated_data):
         try:
-            skills = Skills.objects.all().filter(user_id=self.id)
-            if skills.__len__() > 0:
-                self.data[self.index]['skills'] = [
-                    skill.skill_name for skill in skills
-                ]
-        except Skills.DoesNotExist:
+            bio_data = validated_data.pop("bio")
+            if not Bio.objects.update(user_id=instance.id, **bio_data):
+                Bio.objects.create(user_id=instance.id, **bio_data)
+        except KeyError:
             pass
 
-    def get_address(self):
+    def update_name(self, instance, validated_data):
         try:
-            address = Address.objects.get(user_id=self.id)
-            self.data[self.index]['person_info'][
-                'address'] = address.get_full_address
-        except Address.DoesNotExist:
+            name_data = validated_data.pop("name")
+
+            if not FirstNameAndLastName.objects.update(
+                user_id=instance.id, **name_data
+            ):
+                FirstNameAndLastName.objects.create(user_id=instance.id, **name_data)
+        except KeyError:
             pass