+++ ArabClub_APIs/users/views.py
@@ -1,5 +1,8 @@
 # Url start is "account/"
 from django.contrib.auth import get_user_model
+from django.shortcuts import get_object_or_404
+from rest_framework import status
+from rest_framework.permissions import IsAdminUser
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
@@ -9,71 +12,84 @@ from users.models import (
     Skills,
     Address,
     GitHubAccount,
-    Phone
+    Phone,
 )
 from users.serializers import (
-    CustomSerializerUser,
-    UpdateUserSerializer,
-    UpdateNamesSerializer,
     BioSerializer,
     SkillsSerializer,
     AddressSerializer,
     GitHubSerializer,
-    PhoneSerializer
+    PhoneSerializer,
+    # New
+    UserSerializer,
 )
 
-# from rest_framework.permissions import DjangoModelPermissionsOrAnonReadOnly
 
-
-# TODO
-"""
-Create Login, Register API
-update profile API
-"""
-
-
+# List All User endpoint with custom format
+# Done
 class ListUserView(APIView):
+    permission_classes = [IsAdminUser]
 
     def get(self, request):
         users = get_user_model()
         users = users.objects.all()
-        serializer = CustomSerializerUser(users)
+        serializer = UserSerializer(users, many=True)
         return Response(serializer.data)
 
 
+# View data specific user
 class UserDetailsView(APIView):
+    # permission_classes = [IsOwner]
 
-    @staticmethod
-    def get_object(username):
-        user = get_user_model()
-        try:
-            user = user.objects.get(username=username)
-        except user.DoesNotExist:
-            return ValueError(user.DoesNotExist)
-        return user
+    # Start Old Code
+    # def get_object(self, username):
+    #     user = get_user_model()
+    #     obj = get_object_or_404(user, username=username)
+    #     self.check_object_permissions(self.request, obj)
+    #     return obj
+    #
+    # def get(self, request, username):
+    #     user = self.get_object(username)
+    #     serializer = CustomSerializerUser(user)
+    #     if serializer.err:
+    #         return Response(serializer.err)
+    #     return Response(serializer.data)
+    #
+    # def put(self, request, username):
+    #     serializer = UpdateUserSerializer(data=request.data)
+    #     if not serializer.is_valid():
+    #         return Response(serializer.errors)
+    #     user = self.get_object(username)
+    #     serializer.update(user, serializer.validated_data)
+    #     return Response(serializer.data)
+    # End Old Code
 
-    @staticmethod
-    def get_user_object(request):
-        obj = get_user_model()
-        obj = obj.objects.get(pk=request.user.pk)
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.username = None
+
+    def get_object(self):
+        model = get_user_model()
+        obj = get_object_or_404(model, username=self.username)
         return obj
 
     def get(self, request, username):
-        user = self.get_object(username)
-        serializer = CustomSerializerUser(user)
-        if serializer.err:
-            return Response(serializer.err)
+        self.username = username
+        obj = self.get_object()
+        serializer = UserSerializer(obj, context={"request": request})
         return Response(serializer.data)
 
     def put(self, request, username):
-        serializer = UpdateUserSerializer(data=request.data)
-        if not serializer.is_valid():
-            return Response(serializer.errors)
-        user = self.get_user_object(request)
-        serializer.update(user, serializer.validated_data)
-        return Response(serializer.data)
+        serializer = UserSerializer(data=request.data)
+        if serializer.is_valid():
+            user = get_user_model()
+            user = user.objects.get(pk=request.user.id)
+            serializer.update(user, serializer.validated_data)
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
+# view person info for specific user
 class PersonInfoView(APIView):
     def get_object(self):
         try:
@@ -82,13 +98,13 @@ class PersonInfoView(APIView):
         except FirstNameAndLastName.DoesNotExist:
             pass
 
-    def get(self, request, username):
+    def get(self, request, **kwargs):
         obj = self.get_object()
         serializer = UpdateNamesSerializer(obj)
         return Response(serializer.data)
 
-    def put(self, request, username):
-        self.request.data['user_id'] = request.user.id
+    def put(self, request, **kwargs):
+        self.request.data["user_id"] = request.user.id
         serializer = UpdateNamesSerializer(data=request.data)
         if not serializer.is_valid():
             return Response(serializer.errors)
@@ -97,6 +113,7 @@ class PersonInfoView(APIView):
         return Response(serializer.data)
 
 
+# view bio for specific user
 class BioView(APIView):
     def get_object(self):
         try:
@@ -105,13 +122,13 @@ class BioView(APIView):
         except Bio.DoesNotExist:
             pass
 
-    def get(self, request, username):
+    def get(self, request, **kwargs):
         bio = self.get_object()
         serializer = BioSerializer(bio)
         return Response(serializer.data)
 
-    def put(self, request, username):
-        self.request.data['user_id'] = request.user.id
+    def put(self, request, **kwargs):
+        self.request.data["user_id"] = request.user.id
         serializer = BioSerializer(data=request.data)
         if not serializer.is_valid():
             return Response(serializer.errors)
@@ -120,8 +137,8 @@ class BioView(APIView):
         return Response(serializer.data)
 
 
+# view skills for specific user
 class SkillsView(APIView):
-
     def get_object(self):
         try:
             obj = Skills.objects.all().filter(user_id=self.request.user.id)
@@ -129,39 +146,39 @@ class SkillsView(APIView):
         except Skills.DoesNotExist:
             pass
 
-    def get(self, request, username):
+    def get(self, request, **kwargs):
         obj = self.get_object()
         serializer = SkillsSerializer(obj, many=True)
         return Response(serializer.data)
 
-    def put(self, request, username):
+    def put(self, request, **kwargs):
         for skill in request.data:
             serializer = SkillsSerializer(data=skill)
             if not serializer.is_valid():
                 return Response(serializer.errors)
 
             try:
-                obj = Skills.objects.get(id=skill['id'])
+                obj = Skills.objects.get(id=skill["id"])
             except KeyError:
                 Skills.objects.create(**skill)
                 continue
 
-            if obj.skill_name == skill['skill_name']:
+            if obj.skill_name == skill["skill_name"]:
                 continue
             serializer.update(obj, serializer.validated_data)
         serializer = SkillsSerializer(self.get_object(), many=True)
         return Response(serializer.data)
 
 
+# view address for specific user
 class AddressView(APIView):
-
-    def get(self, request, username):
+    def get(self, request, **kwargs):
         obj = Address.objects.get(user_id=request.user.id)
         serializer = AddressSerializer(obj)
         return Response(serializer.data)
 
-    def put(self, request, username):
-        request.data['user_id'] = request.user.id
+    def put(self, request, **kwargs):
+        request.data["user_id"] = request.user.id
         serializer = AddressSerializer(data=request.data)
         if not serializer.is_valid():
             return Response(serializer.errors)
@@ -170,14 +187,15 @@ class AddressView(APIView):
         return Response(serializer.data)
 
 
+# view github account for specific user
 class GitHubView(APIView):
-    def get(self, request, username):
+    def get(self, request, **kwargs):
         obj = GitHubAccount.objects.get(user_id=request.user.id)
         serializer = GitHubSerializer(obj)
         return Response(serializer.data)
 
-    def put(self, request, username):
-        request.data['user_id'] = request.user.id
+    def put(self, request, **kwargs):
+        request.data["user_id"] = request.user.id
         serializer = GitHubSerializer(data=request.data)
         if not serializer.is_valid():
             return Response(serializer.errors)
@@ -186,14 +204,15 @@ class GitHubView(APIView):
         return Response(serializer.data)
 
 
+# view phone number for specific user
 class PhoneView(APIView):
-    def get(self, request, username):
+    def get(self, request, **kwargs):
         obj = Phone.objects.get(user_id=request.user.id)
         serializer = PhoneSerializer(obj)
         return Response(serializer.data)
 
-    def put(self, request, username):
-        request.data['user_id'] = request.user.id
+    def put(self, request, **kwargs):
+        request.data["user_id"] = request.user.id
         serializer = PhoneSerializer(data=request.data)
         if not serializer.is_valid():
             return Response(serializer.errors)